IA_Generativa/Projeto2/requirements.txt
google-genai
python-dotenv
gradio
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
pypdf
sentence-transformers
langchain-huggingface
mcp
edge-tts